requirements.txt
asgiref==3.5.2
Django==4.1.1
django-crispy-forms==1.14.0
filetype==1.1.0
mysql-connector==2.2.9
Pillow==9.2.0
PyMySQL==1.0.2
sqlparse==0.4.3
tzdata==2022.4
